refactor: ViewController의 createAlert 이동, presentAlert 네이밍 변경
createAlert를 presentAlert로 네이밍 변경
ViewController에 있던 presentAlert을 AlertSetting에서 타입 메서드로 이동 및 변경
